Reduce code Warnings (#80)
* Reduce type convert warning

* Reduce more type convert warnings

* Still reduce type convert warnings

* Remove all type warnings in Engine

* Remove all type warnings in Breakout

* Remove all type warnings in Snake

+++ Learning2DEngine/Learning2DEngine/DebugTool/DebugBoxColliderRenderer.h
@@ -25,7 +25,7 @@ namespace Learning2DEngine
 			GLuint vao;
 			GLuint vboStatic;
 			GLuint vboDynamic;
-			unsigned int maxObjectSize;
+			size_t maxObjectSize;
 
 			std::map<int, std::vector<DebugRenderData<Physics::BaseBoxColliderComponent>*>> debugRenderData;
 			Render::BaseColorDynamicData* dynamicData;
